+++ docs/fr_FR/changelog.md
@@ -4,6 +4,8 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte
 
+- Ajout de parametre de delta consigne - temperature exterieure pour le calcul de lma direction chaud/froid
+
 # 07/04/2020
 
 - Correction de bugs